+++ test/event_utils.h
@@ -7,7 +7,7 @@
 
 namespace hhal {
     namespace events{
-        void write(HHAL &hhal, int event_id, uint8_t value);
+        void write(HHAL &hhal, int event_id, uint32_t value);
         uint32_t read(HHAL &hhal, int event_id);
         uint32_t lock(HHAL &hhal, int event_id);
         void wait(HHAL &hhal, int event_id, uint32_t state);
